Add read_uuid option
In some implementations, a BLE UART uses notifications/indications to send data to the recipient as it is available. The "--read-uuid" AKA "-r" option introduced in this commit allows the user to specify a Gatt Characteristic UUID that ble-serial will attempt to subscribe to notifications from by writing the CCCD of that characteristic.

Known limitations:
- This option does not currently allow users to enable indications nor both notifications and indications at the same time.